pyproject.toml
[project]
name = "docpilot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-aiplatform>=1.122.0",
    "httpx>=0.28.1",
    "ipykernel>=7.0.1",
    "jinja2>=3.1.6",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "nbformat>=5.10.4",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "polars>=1.34.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pypdf>=6.1.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest-cov>=7.0.0",
]
